requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2-binary==2.8.5
python-dotenv==0.13.0
django-admin-json-editor==0.2.0
